% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btw2021-package.R
\docType{data}
\name{afd2021}
\alias{afd2021}
\title{AfD 2021 Election Manifesto}
\format{
An object of class \code{character} of length 199.
}
\usage{
afd2021
}
\description{
The data has been generated with pdftools. URLs have been
researched by the greatest hackers in the world.
}
\details{
See the script in the data raw folder.
}
\keyword{datasets}
