% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btw2021-package.R
\docType{package}
\name{btw2021}
\alias{btw2021}
\title{Manifestos for the 2021 Federal Election in Germany}
\description{
This package includes the manifestos for the 2021 federal election.
}
\author{
Andreas Blaette
}
