% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btw2021-package.R
\docType{data}
\name{cdu_csu2021}
\alias{cdu_csu2021}
\title{CDU/CSU 2021 Election Manifesto}
\format{
An object of class \code{character} of length 140.
}
\usage{
cdu_csu2021
}
\description{
CDU/CSU 2021 Election Manifesto
}
\keyword{datasets}
