% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btw2021-package.R
\docType{data}
\name{fdp2021}
\alias{fdp2021}
\title{FDP 2021 Election Manifesto}
\format{
An object of class \code{character} of length 91.
}
\usage{
fdp2021
}
\description{
FDP 2021 Election Manifesto
}
\keyword{datasets}
