% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btw2021-package.R
\docType{data}
\name{gruene2021}
\alias{gruene2021}
\title{B端ndnis 90/Die Gr端nen 2021 Election Manifesto}
\format{
An object of class \code{character} of length 135.
}
\usage{
gruene2021
}
\description{
B端ndnis 90/Die Gr端nen 2021 Election Manifesto
}
\keyword{datasets}
