% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btw2021-package.R
\docType{data}
\name{linke2021}
\alias{linke2021}
\title{Die LINKE 2021 Election Manifesto}
\format{
An object of class \code{character} of length 148.
}
\usage{
linke2021
}
\description{
Die LINKE 2021 Election Manifesto
}
\keyword{datasets}
