% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btw2021-package.R
\docType{data}
\name{manifestos2021}
\alias{manifestos2021}
\title{List of text content}
\format{
An object of class \code{list} of length 6.
}
\usage{
manifestos2021
}
\description{
The data has been generated with pdftools. URLs have been
researched by the greatest hackers in the world.
}
\keyword{datasets}
