% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btw2021-package.R
\docType{data}
\name{spd2021}
\alias{spd2021}
\title{SPD 2021 Election Manifesto}
\format{
An object of class \code{character} of length 66.
}
\usage{
spd2021
}
\description{
SPD 2021 Election Manifesto
}
\keyword{datasets}
